+++ packages/sit-onyx/src/components/OnyxIconButton/OnyxIconButton.ct.tsx
@@ -1,6 +1,53 @@
-import { test } from "../../playwright-axe";
+import { expect, test } from "../../playwright-axe";
 import { createScreenshotsForAllStates, mockPlaywrightIcon } from "../../utils/playwright";
 import OnyxIconButton from "./OnyxIconButton.vue";
+import type { OnyxIconButtonProps } from "./types";
+
+test("should behave correctly", async ({ mount }) => {
+  let clicks = 0;
+  const setup = {
+    props: {
+      label: "trigger something",
+      icon: mockPlaywrightIcon,
+    } satisfies OnyxIconButtonProps,
+    on: {
+      click: () => clicks++,
+    },
+  };
+
+  // ARRANGE
+  const component = await mount(OnyxIconButton, setup);
+
+  await test.step("clickable by default", async () => {
+    // ACT
+    await component.click();
+    // ASSERT
+    expect(clicks).toBe(1);
+  });
+
+  await test.step("not interactive when disabled ", async () => {
+    // ARRANGE
+    await component.update({ ...setup, props: { disabled: true } });
+    // ASSERT
+    await expect(component).toBeDisabled();
+  });
+
+  await test.step("not interactive when loading ", async () => {
+    // ARRANGE
+    await component.update({ ...setup, props: { disabled: false, loading: true } });
+    // ASSERT
+    await expect(component).toBeDisabled();
+  });
+
+  await test.step("clickable again ", async () => {
+    // ARRANGE
+    await component.update({ ...setup, props: { disabled: false, loading: false } });
+    // ACT
+    await component.click();
+    // ASSERT
+    expect(clicks).toBe(2);
+  });
+});
 
 const STATES = {
   state: ["default", "disabled", "loading"],