+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.ct.tsx
@@ -82,16 +82,43 @@ test("should show optional marker", async ({ mount }) => {
 });
 
 const STATES = {
-  variant: ["default", "placeholder", "initialValue", "loading"],
+  variant: ["default", "placeholder", "initialValue", "loading", "autofill"],
   writeMode: ["write", "readonly", "disabled"],
   focusState: ["", "hover", "focus"],
 } as const;
+
+test("should show message", async ({ mount }) => {
+  // ARRANGE
+  const component = await mount(
+    <OnyxInput label="Label" message="Test message" style="width: 12rem;" />,
+  );
+  const input = component.getByLabel("Label");
+
+  // ACT
+  await component.getByText("Test message").focus();
+
+  // ASSERT
+  await expect(component).toContainText("Test message");
+  await expect(input).not.toBeFocused();
+  await expect(component).toHaveScreenshot("message.png");
+});
+
+test("should show counter", async ({ mount }) => {
+  // ARRANGE
+  const component = await mount(
+    <OnyxInput label="Label" maxlength={16} modelValue="Test" withCounter style="width: 12rem;" />,
+  );
+
+  // ASSERT
+  await expect(component).toContainText("4/16");
+  await expect(component).toHaveScreenshot("counter.png");
+});
 
 test(
   "State screenshot testing",
   createScreenshotsForAllStates(
     STATES,
-    "button",
+    "input",
     async ({ variant, writeMode, focusState }, mount) => {
       const component = await mount(
         <OnyxInput
@@ -101,11 +128,16 @@ test(
           readonly={writeMode === "readonly"}
           disabled={writeMode === "disabled"}
           loading={variant === "loading"}
+          autocomplete={variant === "autofill" ? "name" : undefined}
           style="width: 12rem;"
         />,
       );
 
       const input = component.getByLabel("Label");
+
+      if (variant == "autofill") {
+        await input.evaluate((node) => node.setAttribute("data-test-autofill", ""));
+      }
 
       if (focusState === "hover") await input.hover();
       if (focusState === "focus") await input.focus();