+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # alpha-test-app
 
+## 0.0.1-alpha.42
+
+### Patch Changes
+
+- Updated dependencies [27c2bb8]
+  - sit-onyx@1.0.0-alpha.45
+
+## 0.0.1-alpha.41
+
+### Patch Changes
+
+- Updated dependencies [3ec3d94]
+  - sit-onyx@1.0.0-alpha.44
+
 ## 0.0.1-alpha.40
 
 ### Patch Changes