+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-alpha.45
+
+### Minor Changes
+
+- 27c2bb8: feat(OnyxInput): add styles for autofill state
+
+## 1.0.0-alpha.44
+
+### Minor Changes
+
+- 3ec3d94: feat(OnyxInput): add property `maxlength`, `minLength`, `withCounter` and `message`
+
 ## 1.0.0-alpha.43
 
 ### Minor Changes